ensure tests use installed library
