+++ legateboost/test/models/test_tree.py
@@ -3,9 +3,7 @@ import pytest
 
 import cunumeric as cn
 import legateboost as lb
-
-from ..utils import non_increasing
-from .utils import check_determinism
+from legateboost.testing.utils import check_determinism, non_increasing
 
 
 @pytest.mark.parametrize("max_depth", [0, 8])
