+++ legateboost/test/test_objective.py
@@ -2,8 +2,7 @@ import pytest
 
 import cunumeric as cn
 import legateboost as lb
-
-from .utils import non_increasing
+from legateboost.testing.utils import non_increasing
 
 
 def test_normal() -> None:
