+++ legateboost/testing/__init__.py
@@ -1,0 +1,6 @@
+"""Shared utilities used by legateboost's unit tests.
+
+Not intended to be imported outside of legateboost.
+"""
+
+__all__: list[str] = []
